onboarding (#6)
This pull request contains changes related to a new feature that implements onboarding. A new onboarding screen and its widgets have been added. Additionally, the screen has been registered in AutoRoute, and a Guard has been created to check if onboarding was completed before transitioning to the main screen. If not, it redirects to the onboarding screen. Furthermore, improvements have been made to the application's initialization system.